% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_data.R
\name{addMissingVals}
\alias{addMissingVals}
\title{Add clones that haven't been counted
If clones haven't been counted they don't appear rather than
count as zero}
\usage{
addMissingVals(data_Nths, frac_size)
}
\description{
Add clones that haven't been counted
If clones haven't been counted they don't appear rather than
count as zero
}

