% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_clean1.R
\name{dr_clean1}
\alias{dr_clean1}
\title{One-point drift correction}
\usage{
dr_clean1(dataFrame, varName, calVal, calStd, correctVar)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{varName}{The name of the variable to correct}

\item{calVal}{A number}

\item{calStd}{A number}

\item{correctVar}{Name of value or variable generated from \code{\link{dr_correct()}}}
}
\value{
A list of values for the specified \code{varName} corrected for drift
}
\description{
One-point drift correction
}
\examples{
dr_clean1(df, SpCond, 1.05, 1, corrfactors)
dr_clean1(df, DO, 96.4, 99, df$corrections)

}
