% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_clean2.R
\name{dr_clean2}
\alias{dr_clean2}
\title{Two-point drift correction}
\usage{
dr_clean2(dataFrame, varName, calValLow, calStdLow, calValHigh, calStdHigh,
  correctVar)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{varName}{The name of the variable to correct}

\item{calValLow}{The number that the instrument was actually reading for the low standard}

\item{calStdLow}{The number that the instrument should have been reading for that standard; i.e. the low standard value}

\item{calValHigh}{The number that the instrument was actually reading for the high standard}

\item{calStdHigh}{The number that the instrument should have been reading for that standard; i.e. the high standard value}

\item{correctVar}{Name of value or variable generated from \code{\link{dr_correct}}}
}
\value{
A list of values for the specified \code{varName} corrected for drift
}
\description{
This command takes the raw data from the water-quality instrument, utilizes the values generated from \code{\link{dr_correct}} and returns data that accounts for drift over time. This is done via a two point calibration standard, which it typical for pH and chloride.
}
\examples{
\dontrun{
dr_clean2(df, pH, 7.05, 7, 10.25, 10, corrfactors)
dr_clean2(df, Chloride, 7.95, 10, 847, 1000, df$corrections)
}

}
