% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_correct.R
\name{dr_correct}
\alias{dr_correct}
\title{Creating correction factors}
\usage{
dr_correct(dataFrame, dateVar, timeVar, format)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{dateVar}{Date variable name}

\item{timeVar}{Time variable name}

\item{format}{Either "MDY" or "YMD" for \code{dateVar}}
}
\value{
A series of correction factors
}
\description{
This function creates a series of correction factors. These are calculated based on the time of the observation and the total amount of time that the instrument had been deployed
}
\examples{
\dontrun{
dr_correct(df, Date, Time, "YMD")
dr_correct(df, X, Y, "MDY")
}

}
