% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_correctOne.R
\name{dr_correctOne}
\alias{dr_correctOne}
\title{One-point drift correction}
\usage{
dr_correctOne(.data, sourceVar, cleanVar, calVal, calStd, factorVar)
}
\arguments{
\item{.data}{The working data frame}

\item{sourceVar}{The name of the variable to correct}

\item{cleanVar}{The new var}

\item{calVal}{The number that the instrument was actually reading for the parameter}

\item{calStd}{The number that the instrument should have been reading for that standard; i.e. the standard value}

\item{factorVar}{Name of value or variable generated from \code{\link{dr_factor}}}
}
\value{
A list of values for the specified \code{varName} corrected for drift
}
\description{
This command takes the raw data from the water-quality instrument, utilizes the values generated from \code{\link{dr_factor}} and returns data that accounts for drift over time. This is done via a one point calibration standard, which it typical for specific conductivity, dissolved oxygen, and turbidity.
}
\examples{
\dontrun{
dr_correctOne(df, SpCond, 1.05, 1, corrfactors)
dr_correctOne(df, DO, 96.4, 99, df$corrections)
}

}
