% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_readHOBO.R
\name{dr_readHOBO}
\alias{dr_readHOBO}
\title{Import raw data from an Onset Computer Corporation U24 Conductivity Logger}
\usage{
dr_readHOBO(file, defineVar = TRUE, fileFormat = c("TXT", "CSV"))
}
\arguments{
\item{file}{The name of the file which the data are to be read from. Each row of the table appears
as one line of the file. If it does not contain an absolute path, the file name is relative to
the current working directory.}

\item{defineVar}{Logical scalar that determines if the units of measurement are included in the first
observation. If they are included, all vectors will be read in as character.}

\item{fileFormat}{The format of the input file. HOBO data can be exported as both .csv and .txt files from the instrument.}
}
\value{
A tibble with the formatted data and the variable types defined if \code{defineVar = TRUE}
}
\description{
This function imports the raw data from an Onset U24 Conductivity Logger and formats the data set as a tibble.
    If \code{defineVar} is set to \code{TRUE} (the default option), units of measurement will not be
    included in the first observation.
}
\examples{
\dontrun{
dr_readHOBO("data.txt", defineVar = TRUE, fileFormat = "TXT")
dr_readHOBO("data.csv", defineVar = TRUE, fileFormat = "CSV")
}

}
