% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_exp_data}
\alias{format_exp_data}
\title{Calculate proportions with 95\% credible interval (confidence interval).
If there are replicates they will be pooled(!). It will work for both count table format and list format.}
\usage{
format_exp_data(data_x, time_points = NULL, clone_fractions = NULL)
}
\description{
Calculate proportions with 95\% credible interval (confidence interval).
If there are replicates they will be pooled(!). It will work for both count table format and list format.
}

