% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_list2tbl}
\alias{format_list2tbl}
\title{Transform list format to table format with clone size along rows and time points
along cols. Will return time points as colnames. This is the format used for inference
If a clone size does not appear in any data point it will not be included (!!!)}
\usage{
format_list2tbl(list_x, condition_filter, clone_fractions = NULL)
}
\description{
Transform list format to table format with clone size along rows and time points
along cols. Will return time points as colnames. This is the format used for inference
If a clone size does not appear in any data point it will not be included (!!!)
}

