% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_tbl2list}
\alias{format_tbl2list}
\title{Transform tbl format to list format. Useful for ggplot plotting}
\usage{
format_tbl2list(tbl_x, time_points, condition = "Condition1")
}
\description{
Transform tbl format to list format. Useful for ggplot plotting
}

