% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notInOperator.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In Operator}
\source{
\href{\code{Hmsic}}{https://github.com/harrelfe/Hmisc/blob/master/R/in.operator.s}
}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched}

\item{table}{vector or \code{NULL}: the values to be matched against}
}
\description{
Provides the compliment to the base R \code{\%in\%} operator. Included here instead of via import
due to stability issues with the source package, \href{\code{Hmsic}}{https://github.com/harrelfe/Hmisc/blob/master/R/in.operator.s},
during original package development in October, 2017. Used under terms of \href{\code{Hmisc}}{https://cran.r-project.org/web/packages/Hmisc/index.html}'s
\href{GPL-3 License}{https://cran.r-project.org/web/licenses/GPL-3}.
}
\examples{
x <- 2
y <- 2
z <- 3

x \%in\% y
x \%nin\% y

x \%in\% z
x \%nin\% z

}
