% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_clean1.R
\name{dr_clean1}
\alias{dr_clean1}
\title{One-point drift correction}
\usage{
dr_clean1(dataFrame, varName, calVal, calStd, correctVar)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{varName}{The name of the variable to correct}

\item{calVal}{The number that the instrument was actually reading for the parameter}

\item{calStd}{The number that the instrument should have been reading for that standard; i.e. the standard value}

\item{correctVar}{Name of value or variable generated from \code{\link{dr_correct}}}
}
\value{
A list of values for the specified \code{varName} corrected for drift
}
\description{
This command takes the raw data from the water-quality instrument, utilizes the values generated from \code{\link{dr_correct}} and returns data that accounts for drift over time. This is done via a one point calibration standard, which it typical for specific conductivity, dissolved oxygen, and turbidity.
}
\examples{
\dontrun{
dr_clean1(df, SpCond, 1.05, 1, corrfactors)
dr_clean1(df, DO, 96.4, 99, df$corrections)
}

}
