% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_clean2.R
\name{dr_clean2}
\alias{dr_clean2}
\title{Two-point drift correction}
\usage{
dr_clean2(dataFrame, varName, calValLow, calStdLow, calValHigh, calStdHigh,
  correctVar)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{varName}{The name of the variable to correct}

\item{calValLow}{A number}

\item{calStdLow}{A number}

\item{calValHigh}{A number}

\item{calStdHigh}{A number}

\item{correctVar}{Name of value or variable generated from \code{\link{dr_correct}}}
}
\value{
A list of values for the specified \code{varName} corrected for drift
}
\description{
Two-point drift correction
}
\examples{
\dontrun{
dr_clean2(df, pH, 7.05, 7, 10.25, 10, corrfactors)
dr_clean2(df, Chloride, 7.95, 10, 847, 1000, df$corrections)
}

}
