% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_correctTwo.R
\name{dr_correctTwo}
\alias{dr_correctTwo}
\title{Two-point drift correction}
\usage{
dr_correctTwo(.data, sourceVar, cleanVar, calValLow, calStdLow, calValHigh,
  calStdHigh, factorVar)
}
\arguments{
\item{.data}{A tbl}

\item{sourceVar}{Name of variable to correct}

\item{cleanVar}{New variable name for corrected data}

\item{calValLow}{A numeric value; the number that the instrument was actually reading for
the low standard}

\item{calStdLow}{A numeric value; the number that the instrument should have been reading
for that standard; i.e. the low standard value}

\item{calValHigh}{A numeric value; the number that the instrument was actually reading
for the high standard}

\item{calStdHigh}{A numeric value; the number that the instrument should have been reading
for that standard; i.e. the high standard value}

\item{factorVar}{Name of variable generated using \code{\link{dr_factor}}}
}
\value{
An object of the same class as \code{.data} with the new corrected variable added
to the other data in \code{.data}.
}
\description{
A wrapper around \code{dplyr::mutate()} that creates a corrected value for each observation of the
    specified variable based on two data points.
}
\details{
This command takes the raw data from the water-quality instrument, utilizes the values
   generated from \code{\link{dr_factor}} and returns data that accounts for drift over time.
   This is done via a two-point calibration standard, which it typical for pH and chloride.
}
\examples{
\dontrun{
dr_correctTwo(df, pH, 7.05, 7, 10.25, 10, corrfactors)
dr_correctTwo(df, Chloride, 7.95, 10, 847, 1000, df$corrections)
}

}
\seealso{
\code{\link{dr_factor}} for correction factor creation,
    \code{\link{dr_correctOne}} for the two-point drift correction
}
