% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_drop.R
\name{dr_drop}
\alias{dr_drop}
\title{Dropping observations for equilibration period}
\usage{
dr_drop(dataFrame, head = NULL, tail = NULL)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{head}{A number}

\item{tail}{A number}
}
\value{
A data frame with fewer observations
}
\description{
Dropping observations for equilibration period
}
\examples{
\dontrun{
dr_drop(df, 3, 9)
dr_drop(df, , 4)
dr_drop(df, 7,)
}

}
