% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_drop.R
\name{dr_drop}
\alias{dr_drop}
\title{Dropping observations for equilibration period}
\usage{
dr_drop(dataFrame, head = NULL, tail = NULL)
}
\arguments{
\item{dataFrame}{The working data frame}

\item{head}{A number}

\item{tail}{A number}
}
\value{
A data frame with fewer observations
}
\description{
When taking the instrument out of the water, there is often several observations that pass before the run can be downloaded. Additionally, once the instrument is in the water, it often takes about 30 minutes for the sensors to equilibrate. This function allows you to drop observations from the bottom and top of the dataset for each of those issues respectively.
}
\examples{
\dontrun{
dr_drop(df, 3, 9)
dr_drop(df, , 4)
dr_drop(df, 7,)
}

}
