% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_readSonde.R
\name{dr_readSonde}
\alias{dr_readSonde}
\title{Import raw data}
\usage{
dr_readSonde(fileExtention, defineVar = FALSE)
}
\arguments{
\item{fileExtention}{Location of data}

\item{defineVar}{Logical statement}
}
\value{
A dataframe with with the raw data and the variable types defined if \code{defineVar = TRUE}
}
\description{
Import raw data
}
\examples{
dr_readSonde("~/fileLocation/data.csv")
dr_readSonde("~/fileLocation/data.csv", defineVar= TRUE)
}
